Big update

+++ CodeforcesCSharpApp/Ozon/Route256/Sandbox-2022.08.17/ProblemH/Solution-01.cs
@@ -12,7 +12,7 @@ public static class Program
 
         var processors = Console.ReadLine()!
             .Split(' ', StringSplitOptions.RemoveEmptyEntries)
-            .Select(a => new Processor(currentTime, Convert.ToInt32(a)))
+            .Select(a => new Processor(currentTime, Convert.ToInt64(a)))
             .ToArray();
 
         var freeProcessors = new PriorityQueue<Processor, long>();
